E commerce Data Analysis Project.
Analyzing and cleaning the data using Jupyter Notebooks. Exporting as a csv file and then utilizing Power Bi to create an interactive Dashboard to finalize this report for an e commerce website to find popular states in which products are sold more so that advertising efforts can be increased in those specific areas.