Beauty Brand E commerce Data Work
I worked with a large data set of 10,000+ products sold. I analyzed the raw data first and then with the help of Python, I cleaned the csv file for any inconsistencies. Later on, I filtered the data for products which had only 1000+ reviews, gender as for both male and female, and with an average rating of more than 4 stars. Lastly, I worked on Data Visualization with Power Bi and made the final interactive dashboard which can easily be used by any user to see the patterns in terms of revenue generated by each specific product under a specific category in the form of a pie chart as well as display name and the ratings as well as the total number of reviews , over the period of this dataset generation.